requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
prometheus_client==0.11.0
boto3==1.17.72
Werkzeug==2.2.2